Webview: extract rendered event block (oh-tab-ihgw)
